requirements.txt
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
PyYAML==6.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
urllib3==1.26.16
wsproto==1.2.0
zipp==3.15.0
parameterized==0.9.0
requests~=2.31.0
pillow~=10.2.0
opencv-python~=4.9.0.80